// src/types.ts
export interface ShortData {
    id: number;
    name: string;
    longname: string;
    orgname?: string;
    orgid?: number;
}

export interface Lesson {
    id: number;
    date: number;
    startTime: number;
    endTime: number;
    kl: ShortData[];
    te: ShortData[];
    su: ShortData[];
    ro: ShortData[];
    lstext?: string;
    lsnumber: number;
    activityType?: 'Unterricht' | string;
    lessonType?: 'Unterricht' | string;
    code?: 'cancelled' | 'irregular';
    info?: string;
    substText?: string;
    statflags?: string;
    sg?: string;
    bkRemark?: string;
    bkText?: string;
    subject?: string;
}

export interface Homeworks {
    homeworks: Homework[];
    lessons: Partial<Lesson>[];
    records: HomeworkRecord[];
    teachers: Partial<Teacher>[];
}

export interface Homework {
    attachments: any[];
    completed: boolean;
    date: number;
    dueDate: number;
    id: number;
    lessonId: number;
    remark: string;
    text: string;
}

export interface HomeworkRecord {
    elementIds: number[];
    homeworkId: number;
    teacherId: number;
}
// ...
export interface Exam {
    id: number;
    examType: string;
    name: string;
    studentClass: string[];
    assignedStudents: {
        klasse: { id: number; name: string };
        displayName: string;
        id: number;
    }[];
    examDate: number;
    startTime: number;
    endTime: number;
    subject: string;
    teachers: string[];
    rooms: string[];
    text: string;
    grade?: string;
}

export enum WebUntisElementType {
    CLASS = 1,
    TEACHER = 2,
    SUBJECT = 3,
    ROOM = 4,
    STUDENT = 5,
}
// ...
export interface Teacher {
    id: number;
    name: string;
    foreName: string;
    longName: string;
    foreColor?: string;
    backColor?: string;
    active?: boolean;
    dids?: any[];
    title?: string;
}